requirements.txt
numpy
pillow
torch
torchvision
opencv-python
einops
